Разработал редьюсер Sportsman. Доработал компонент отображения общих результатов соревнования.

+++ Models/Result.cs
@@ -3,7 +3,7 @@ namespace circle_competitions_monitoring.Models {
         public int id { get; set; }
         public int sportsman { get; set; }
         public int competition { get; set; }
-        public float points { get; set; }
+        public double points { get; set; }
         public int place { get; set; }
     }
 }
